refactor : Member auth API
